Fix tig blame "open file on GitHub" command
tig's %(file) variable is populated with each row's
filename at commit time, which is undesirable for
this key-binding.

This is fixed with a script that takes a file path
as its only argument, and prints the current file path
of that file, by stepping through its git renaming
history, in reverse.