Add more surround characters, and switch key-combo
`gs` and `gsd` are easier to hit, when the character to use in the
surrounding uses the shift key.

`S` is a synonym for `cc` which is annoying when the timeout on the
key press is reached. `gs` is a vim built-in for `goto sleep`, so the
timeout just makes vim go to sleep for a second.