Rewrite comment handling on pull-requests
The /reviews route returns ALL comments, but the body property is the
empty string on reviews.

The /comments route returns the review comments with the body property
set, but not comments left outside of reviews.

🙄 😵 🤔

So we need to fetch both arrays and merge them, and then sort them by
date, to get them in chronological order.

There is no connection between the comments and the reviews, so we need to
aggregate them by commentId (which is the hunk_id) for the reviews. They
are sorted by date, so the first with a commentId is the parent. All other
comments with the same commentId are added as children.