Translate "docker compose" to "docker-compose"
The Homebrew version of `docker` does not support the `-d` (detatch)
flag for the `compose` command. The separately installed binary
`docker-compose` supports it, however.

This solution intercepts the execution and `compose` commands are sent
to the separate binary.

There is a new directory `~/bin` that is added to the head of the
`$PATH` to make sure it is always executed first.