Include SSH config file in ~/.ssh/config
Since other programs (Colima, for instance) makes edits to this file, it
is easier to include the dotfiles config file, instead of linking it.

If it was linked, changes by other programs would result in a diff in
git. It may also overwrite existing configuration, if run after changes
has been made.

Use vim to add the line to the end of the file, to make it POSIX compliant,
since it may lack a trailing newline. I'm looking at you, Colima...