Vim command to open a JavaScript stack trace in quickfix
Copying the output of a stack trace to the clipboard lets this command
add all file locations to the quickfix list, to easily stepping through
the execution path.

If the first line is the error message, it is used as the title of the
quickfix window. Otherwise, "Stack Trace" will be used.